Ensures fakes for typed Funcs have correct package selectors
The usage of "prefixTypes" was strange and scary, since it clearly is mutating
these nodes. Perhaps changing this code to not mutate the pointers and instead
return the new version of them would be better.

Testing this code, and moving it out of the "locator" package would be a good idea.
This feels like a general purpose transform that should be done in a decorator style,
perhaps? It's certainly not the responsibility of a "locator".

Fixes #29