Ensure Invocations() is race-free
I think the vast amount of mutexes present in some of these generated methods
is probably a code smell. One possible option is that we should limit ourselves
to a single mutex here, so that it's easier to verify by hand that these fakes
are not susceptible to data races. I had thought briefly about this a while back,
but it wasn't until Barrett submitted a fix for this that I realized how broken
this was at the time.

Refactoring this code to generate (and use) a single mutex will cause a lot more
contention in the case that multiple goroutines are calling functions on a fake,
but (in my experience at least) that's a rather unusual edge case.