Update examples in help output
Turns out I wrote this incorrectly almost exactly
one year ago. We should probably strive to rewrite counterfeiter to
leverage a nicer framework for its command and flags so that we
do not have these kinds of embarassing failures when people are
trying to use the features we added to counterfeiter.