Fix race condition in invocation recording
Since we return a map for invocations, which is implemented internally
by a pointer type in golang, we have a very, very interesting data race
here because, while we acquire locks for the process of reading from the
map, once we return it those locks are returned. At that point in time, various
goroutines can possibly try to read or write from the map (which is all bad).

e.g.: the main test goroutine may have called fake.Invocations() and then asserts on it
      a secondary test goroutine is calling a method on the fake, which writes into invocations

So, as a side effect of this commit

* fixed a possible race condition in user tests
* brought back the race flag for counterfeiter's test suite (yay!)
* fixed #59