Return errors instead of os.Exit
os.Exit should only be called from the main() function. If it is
called from other functions then defers will not run before the
program exits.

Also use a string for cwd instead of a function. The value should
not change, so there is no need to make the value lazy.