Allow adding headers to generated fakes
When the `-header` flag is specified and is a valid file path, the
content of that file will be prepended to the generated fake. This is
useful for e.g. specify a licence header.

The header flag can be specified in the `go:generate` and the
`counterfeiter:generate` line or in both, where the later takes
precedence over the former.