Handle methods and functions with generic params
This only covers the case where the interface itself isn't generic, but
accepts or returns a generic with a concret type parameter.

Currently, the fake fails to import the package for the type arg. Now
while creating imports we iterate over TypeArgs for named types and
import the packages for them.